Points mechanicsw
It should work now

+++ Assets/Scripts/BallScore.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class BallScore : MonoBehaviour
 {
-    public float score = 0;
+    public int score = 0;
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
     {
